Move printed files to a new folder
To allow the script to be run multiple times without printing the same 
files over again, the script sends printed files to a new folder titled 
'printed_<today's date>' within the same directory as the files. A 10s 
timer delays the movement of files after each print to ensure that files 
are not moved before they've had a chance to be printed.